requirements.txt
anthropic>=0.49.0
alpaca-py>=0.35.0
APScheduler>=3.10.4
pandas>=2.2.0
praw>=7.8.1
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
SQLAlchemy>=2.0.30
yfinance>=0.2.54
pytrends-modern>=0.2.5
